requirements.txt
PyGithub
Twython
